Pull out types
